Rename OSX to macOS.
